turned Footer into a react component made appropriate Container to allow for adding of version number #89
